Move EncString into its own files
