+++ GWToolboxdll/Utils/GuiUtils.h
@@ -5,9 +5,8 @@
 #include <nlohmann/json.hpp>
 #include <ToolboxIni.h>
 
-namespace GW::Constants {
-    enum class Language;
-}
+// ReSharper disable once CppUnusedIncludeDirective
+#include <Utils/EncString.h>
 
 struct RectF {
     ImVec2 top_left;
@@ -140,57 +139,6 @@ namespace GuiUtils {
     std::string format(const char* msg, ...);
     // Same as std::format, but use printf formatting
     std::wstring format(const wchar_t* msg, ...);
-
-    class EncString {
-    protected:
-        std::wstring encoded_ws;
-        std::wstring decoded_ws;
-        std::string decoded_s;
-        bool decoding = false;
-        bool decoded = false;
-        bool sanitised = false;
-        bool release = false;
-        virtual void sanitise();
-        virtual void decode();
-        GW::Constants::Language language_id = static_cast<GW::Constants::Language>(0xff);
-        static void OnStringDecoded(void* param, const wchar_t* decoded);
-
-    public:
-        // Set the language for decoding this encoded string. If the language has changed, resets the decoded result. Returns this for chaining.
-        EncString* language(GW::Constants::Language l);
-        bool IsDecoding() const { return decoding && decoded_ws.empty(); };
-        // Recycle this EncString by passing a new encoded string id to decode.
-        // Set sanitise to true to automatically remove guild tags etc from the string
-        EncString* reset(uint32_t _enc_string_id = 0, bool sanitise = true);
-        // Recycle this EncString by passing a new string to decode.
-        // Set sanitise to true to automatically remove guild tags etc from the string
-        EncString* reset(const wchar_t* _enc_string = nullptr, bool sanitise = true);
-        std::wstring& wstring();
-        std::string& string();
-
-        // Free memory used by this EncString. 
-        void Release();
-
-        [[nodiscard]] const std::wstring& encoded() const
-        {
-            return encoded_ws;
-        };
-
-        EncString(const wchar_t* _enc_string = nullptr, const bool sanitise = true)
-        {
-            reset(_enc_string, sanitise);
-        }
-
-        EncString(const uint32_t _enc_string, const bool sanitise = true)
-        {
-            reset(_enc_string, sanitise);
-        }
-
-        // Disable object copying; decoded_ws is passed to GW by reference and would be bad to do this. Pass by pointer instead.
-        EncString(const EncString& temp_obj) = delete;
-        EncString& operator=(const EncString& temp_obj) = delete;
-        ~EncString();
-    };
 
     // Create an ImGui representation of the skill bar
     void DrawSkillbar(const char* build_code, bool show_attributes = true);
